apps/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "app-backend"
version = "0.1.0"
description = "Orchestrator — FastAPI backend"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "motor>=3.6",
  "firebase-admin>=6.5",
  "langgraph>=0.3",
  "anthropic>=0.40",
  "litellm",
  "mcp",
  "pydantic-settings>=2.0",
  "asyncpg",
  "aiomysql",
  "httpx",
  "jsonschema",
  "python-jose[cryptography]",
  "hvac",
  "redis",
  "rq",
  "resend",
  "stripe",
  "duckdb",
  "nanoid",
  "sentry-sdk>=2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx",
  "ruff>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["app_backend"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
